Abstraction for local disk path allocation.
LocalDirAllocator is used to do a round robin based disk allocation for paths
that need to be created on the local disk. There is a need to inject a different
implementation for some use cases. So an interface called LocalDiskPathAllocator
is provided. The actual implementation can be retrieved using LocalDiskUtil.